+++ src/main/java/insane96mcp/progressivebosses/modules/dragon/feature/RewardFeature.java
@@ -1,4 +1,4 @@
-package insane96mcp.progressivebosses.modules.wither.feature;
+package insane96mcp.progressivebosses.modules.dragon.feature;
 
 import insane96mcp.insanelib.base.Feature;
 import insane96mcp.insanelib.base.Label;
@@ -6,30 +6,28 @@ import insane96mcp.insanelib.base.Module;
 import insane96mcp.progressivebosses.base.Strings;
 import insane96mcp.progressivebosses.classutils.Drop;
 import insane96mcp.progressivebosses.setup.Config;
-import net.minecraft.entity.boss.WitherEntity;
+import net.minecraft.block.Blocks;
+import net.minecraft.entity.boss.dragon.EnderDragonEntity;
+import net.minecraft.entity.item.ExperienceOrbEntity;
 import net.minecraft.nbt.CompoundNBT;
+import net.minecraft.util.math.BlockPos;
 import net.minecraftforge.common.ForgeConfigSpec;
-import net.minecraftforge.event.entity.EntityJoinWorldEvent;
 import net.minecraftforge.event.entity.living.LivingDeathEvent;
+import net.minecraftforge.event.entity.living.LivingEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
-@Label(name = "Rewards", description = "Bonus Experience and Drops")
+@Label(name = "Rewards", description = "Bonus Experience and Dragon Egg per player")
 public class RewardFeature extends Feature {
 
 	private final ForgeConfigSpec.ConfigValue<Double> bonusExperienceConfig;
+	private final ForgeConfigSpec.ConfigValue<Boolean> dragonEggPerPlayerConfig;
 	private final ForgeConfigSpec.ConfigValue<List<? extends String>> dropsListConfig;
 
-	/*private static final List<String> dropsListDefault = Arrays.asList("progressivebosses:nether_star_shard,1,2,8,MINIMUM,SCALING",
-			"progressivebosses:nether_star_shard,2,4,6,MINIMUM,SCALING",
-			"progressivebosses:nether_star_shard,3,6,4,MINIMUM,SCALING",
-			"progressivebosses:nether_star_shard,4,8,2,MINIMUM,SCALING");*/
-	private static final List<String> dropsListDefault = Arrays.asList("progressivebosses:nether_star_shard,1,1,0.16,PER_DIFFICULTY,FLAT");
-
-	public double bonusExperience = 20d;
+	public double bonusExperience = 0.3d;
+	public boolean dragonEggPerPlayer = true;
 	public ArrayList<Drop> dropsList;
 
 	public RewardFeature(Module module) {
@@ -38,6 +36,9 @@ public class RewardFeature extends Feature {
 		bonusExperienceConfig = Config.builder
 				.comment("How much more experience (percentage) will Wither drop per Difficulty. The percentage is additive (e.g. with this set to 10%, 7 withers spawned = 70% more experience)")
 				.defineInRange("Bonus Experience per Difficulty", bonusExperience, 0.0, Double.MAX_VALUE);
+		dragonEggPerPlayerConfig = Config.builder
+				.comment("If true whenever a player, that has never killed the dragon, kills the dragon a Dragon Egg ìì will drop. E.g. If 2 players kill the Dragon for the first time, she will drop 2 Dragon Eggs")
+				.define("Dragon Egg per Player", dragonEggPerPlayer);
 		dropsListConfig = Config.builder
 				.comment("A list of drops for the Withers. Entry format: item,amount,difficulty_required,chance,difficulty_mode,chance_mode\n" +
 						"item: item id\n" +
@@ -51,7 +52,7 @@ public class RewardFeature extends Feature {
 						"* FLAT: chance is the percentage chance for the item to drop if the difficulty criteria matches\n" +
 						"* SCALING: each point of difficulty >= 'difficulty to drop the item' will be multiplied by the chance (e.g. chance 2% and difficulty 10, difficulty required 5, chance to drop the item will be chance * (difficulty - difficulty_required + 1) = 2% * (10 - 5 + 1) = 12%)\n" +
 						"By default Withers have 2% chance per difficulty >= 2 to drop 1 shard + 4% chance per difficulty >= 4 to drop 2 shards + 8% chance per difficulty >= 8 to drop 4 shards.")
-				.defineList("Drops", dropsListDefault, o -> o instanceof String);
+				.defineList("Drops", ArrayList::new, o -> o instanceof String);
 		Config.builder.pop();
 	}
 
@@ -59,29 +60,70 @@ public class RewardFeature extends Feature {
 	public void loadConfig() {
 		super.loadConfig();
 		this.bonusExperience = this.bonusExperienceConfig.get();
+		this.dragonEggPerPlayer = this.dragonEggPerPlayerConfig.get();
 		this.dropsList = Drop.parseDropsList(this.dropsListConfig.get());
 	}
 
 	@SubscribeEvent
-	public void onSpawn(EntityJoinWorldEvent event) {
-		if (event.getWorld().isRemote)
-			return;
-
+	public void onUpdate(LivingEvent.LivingUpdateEvent event) {
 		if (!this.isEnabled())
 			return;
 
+		if (!(event.getEntity() instanceof EnderDragonEntity))
+			return;
+
+		EnderDragonEntity dragon = (EnderDragonEntity) event.getEntity();
+
+		dropExperience(dragon);
+		dropEgg(dragon);
+	}
+
+	private void dropExperience(EnderDragonEntity dragon) {
 		if (this.bonusExperience == 0d)
 			return;
 
-		if (!(event.getEntity() instanceof WitherEntity))
+		if (dragon.deathTicks <= 150 || dragon.deathTicks % 5 != 0)
 			return;
 
-		WitherEntity wither = (WitherEntity) event.getEntity();
+		if (dragon.getFightManager() == null)
+			return;
 
-		CompoundNBT witherTags = wither.getPersistentData();
-		float difficulty = witherTags.getFloat(Strings.Tags.DIFFICULTY);
+		CompoundNBT dragonTags = dragon.getPersistentData();
 
-		wither.experienceValue = 50 + (int) (50 * (this.bonusExperience * difficulty / 100f));
+		float difficulty = dragonTags.getFloat(Strings.Tags.DIFFICULTY);
+		if (difficulty == 0d)
+			return;
+
+		//Drop 8% of experience each tick (for 10 ticks) + 20% at the last tick
+		int bonusXP = (int) (500 * this.bonusExperience * difficulty * 0.08);
+		if (dragon.deathTicks == 200)
+			bonusXP += (int) (500 * this.bonusExperience * difficulty * 0.2);
+
+		while (bonusXP > 0) {
+			int i = ExperienceOrbEntity.getXPSplit(bonusXP);
+			bonusXP -= i;
+			dragon.world.addEntity(new ExperienceOrbEntity(dragon.world, dragon.getPositionVec().getX(), dragon.getPositionVec().getY(), dragon.getPositionVec().getZ(), i));
+		}
+	}
+
+	private void dropEgg(EnderDragonEntity dragon) {
+		if (!this.dragonEggPerPlayer)
+			return;
+
+		if (dragon.deathTicks != 100)
+			return;
+
+		CompoundNBT tags = dragon.getPersistentData();
+
+		int eggsToDrop = tags.getInt(Strings.Tags.EGGS_TO_DROP);
+
+		if (dragon.getFightManager() != null && !dragon.getFightManager().hasPreviouslyKilledDragon()) {
+			eggsToDrop--;
+		}
+
+		for (int i = 0; i < eggsToDrop; i++) {
+			dragon.world.setBlockState(new BlockPos(0, 255 - i, 0), Blocks.DRAGON_EGG.getDefaultState());
+		}
 	}
 
 	@SubscribeEvent
@@ -92,15 +134,16 @@ public class RewardFeature extends Feature {
 		if (this.dropsList.isEmpty())
 			return;
 
-		if (!(event.getEntityLiving() instanceof WitherEntity))
+		if (!(event.getEntityLiving() instanceof EnderDragonEntity))
 			return;
 
-		WitherEntity wither = (WitherEntity) event.getEntityLiving();
+		EnderDragonEntity dragon = (EnderDragonEntity) event.getEntityLiving();
 
-		CompoundNBT tags = wither.getPersistentData();
+		CompoundNBT tags = dragon.getPersistentData();
 		float difficulty = tags.getFloat(Strings.Tags.DIFFICULTY);
 		for (Drop drop : this.dropsList) {
-			drop.drop(wither.world, wither.getPositionVec(), difficulty);
+			drop.drop(dragon.world, dragon.getPositionVec(), difficulty);
 		}
 	}
+
 }